+++ Sources/KeyboardKit/Views/System/SystemKeyboardActionButton.swift
@@ -119,7 +119,7 @@ private extension SystemKeyboardActionButton {
         )
     }
     
-    var buttonStyle: SystemKeyboardButtonStyle {
+    var buttonStyle: KeyboardButtonStyle {
         appearance.buttonStyle(
             for: action,
             isPressed: isPressed)