+++ Sources/KeyboardKit/Views/System/SystemKeyboardButtonBody.swift
@@ -20,11 +20,11 @@ public struct SystemKeyboardButtonBody: View {
      - Parameters:
        - style: The button style to apply.
      */
-    public init(style: SystemKeyboardButtonStyle) {
+    public init(style: KeyboardButtonStyle) {
         self.style = style
     }
     
-    private let style: SystemKeyboardButtonStyle
+    private let style: KeyboardButtonStyle
     
     public var body: some View {
         RoundedRectangle(cornerRadius: style.cornerRadius)